requirements.txt
fastapi
uvicorn
python-multipart
pandas
scikit-learn
nltk
pdfplumber
python-docx
duckduckgo_search
ddgs
primp
